refactor: extract render domain (html/build/serve) into src/render/
