--- src/build.ts
@@ -1,13 +1,0 @@
-import { readCatalog } from "./catalog/read";
-import { emitHtml } from "./renderers/html";
-
-export interface BuildOptions {
-  owlDir: string;
-  outDir: string;
-  webBundleDir?: string;
-}
-
-export const runBuild = ({ owlDir, outDir, webBundleDir }: BuildOptions): void => {
-  const catalog = readCatalog(owlDir);
-  emitHtml(catalog, outDir, webBundleDir);
-};
